주문 Service 구현 완료, OrderPageItemDTO 칼럼수정, main.jsp수정,CartMapper 오류수정

+++ src/main/java/com/ht/order/domain/OrderPageItemDTO.java
@@ -23,7 +23,7 @@ public class OrderPageItemDTO {
 	@Getter@Setter
 	private int bookPrice;
 	@Getter@Setter
-	private int bookDiscount;
+	private double bookDiscount;
 	
 	// 만들어 놓을 값
 	@Getter@Setter
@@ -40,7 +40,7 @@ public class OrderPageItemDTO {
 	private List<AttachImageVO> imageList;
 	
 	public void initSaleTotal() {
-		this.sale_price = (int) (this.bookPrice * (1-this.bookDiscount));
+		this.sale_price = (int)(this.bookPrice * (1-this.bookDiscount));
 		this.total_price = this.sale_price*this.book_count;
 		this.point = (int)(Math.floor(this.sale_price*0.05));
 		this.total_point =this.point * this.book_count;
